Move easing function enum to Enums folder
