+++ HtmlForgeX/Containers/VisNetwork/Enums/VisNetworkEasingFunction.cs
@@ -1,0 +1,33 @@
+namespace HtmlForgeX;
+
+/// <summary>
+/// Easing functions for animations.
+/// </summary>
+public enum VisNetworkEasingFunction {
+    /// <summary>Linear easing.</summary>
+    Linear,
+    /// <summary>Easing that accelerates from zero velocity.</summary>
+    EaseInQuad,
+    /// <summary>Easing that decelerates to zero velocity.</summary>
+    EaseOutQuad,
+    /// <summary>Easing that accelerates then decelerates.</summary>
+    EaseInOutQuad,
+    /// <summary>Cubic ease-in.</summary>
+    EaseInCubic,
+    /// <summary>Cubic ease-out.</summary>
+    EaseOutCubic,
+    /// <summary>Cubic ease-in-out.</summary>
+    EaseInOutCubic,
+    /// <summary>Quartic ease-in.</summary>
+    EaseInQuart,
+    /// <summary>Quartic ease-out.</summary>
+    EaseOutQuart,
+    /// <summary>Quartic ease-in-out.</summary>
+    EaseInOutQuart,
+    /// <summary>Quintic ease-in.</summary>
+    EaseInQuint,
+    /// <summary>Quintic ease-out.</summary>
+    EaseOutQuint,
+    /// <summary>Quintic ease-in-out.</summary>
+    EaseInOutQuint
+}
